@@ -11,18 +11,19 @@
   "license": "ISC",
   "repository": "github:itsarraj/pr-scan-action",
   "dependencies": {
-    "@octokit/rest": "^20.1.1",
-    "@probot/adapter-github-actions": "^3.1.1",
-    "console.table": "^0.10.0"
+    "@octokit/rest": "^21.0.2",
+    "@probot/adapter-github-actions": "^4.0.0",
+    "console.table": "^0.10.0",
+    "node-fetch": "^3.3.2"
   },
   "devDependencies": {
-    "@babel/plugin-transform-class-properties": "^7.22.5",
+    "@babel/plugin-transform-class-properties": "^7.25.7",
     "@semantic-release/git": "^10.0.1",
-    "@vercel/ncc": "^0.36.1",
-    "eslint": "^8.45.0",
-    "nock": "^13.3.2",
-    "semantic-release": "^21.0.7",
-    "typescript": "^5.1.6",
+    "@vercel/ncc": "^0.38.2",
+    "eslint": "^9.13.0",
+    "nock": "^13.5.5",
+    "semantic-release": "^24.1.3",
+    "typescript": "^5.6.3",
     "uvu": "^0.5.6"
   },
   "release": {
