@@ -1,4 +1,7 @@
-const { run } = require("@probot/adapter-github-actions");
-const botApp = require("./app");
+import { run } from "@probot/adapter-github-actions";
+import app from "./app.ts";
 
-run(botApp);
+run(app).catch((error) => {
+  console.error(error);
+  process.exit(1);
+});
